improves structure still more
